requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
DAWG-Python==0.7.2
DAWG2-Python==0.9.0
docopt==0.6.2
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.3.0
gensim==4.4.0
h11==0.16.0
hdbscan==0.8.42
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kafka-python==2.0.2
kafka-python==2.3.0
kiwisolver==1.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
nltk==3.9.4
numpy==2.4.4
packaging==26.0
pandas==3.0.2
requests==2.32.3
psycopg[binary,pool]==3.2.3
pymorphy2==0.9.1
pillow==12.2.0
psycopg==3.3.3
psycopg-pool==3.3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
pymorphy2-dicts-ru==2.4.417127.4579844
pydantic==2.9.2
pydantic-settings==2.5.2
pymorphy3==2.0.6
pymorphy3-dicts-ru==2.4.417150.4580142
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.3.32
requests==2.33.1
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
torch==2.4.1
sympy==1.14.0
threadpoolctl==3.6.0
transformers==4.45.2
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.4.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wrapt==2.1.2
